Start to support other numeric types in branch distance

+++ client-dotnet/test/EvoMaster.Instrumentation.Examples/Numbers/NumericOperations.cs
@@ -1,18 +1,18 @@
 ﻿namespace EvoMaster.Instrumentation.Examples.Numbers {
     public class NumericOperations {
-        // public bool GreaterThan(double a, double b) {
-        //     if (a > b)
-        //         return true;
-        //     else
-        //         return false;
-        // }
-        
-        public bool GreaterThan(int a, int b) {
+        public bool GreaterThan(double a, double b) {
             if (a > b)
                 return true;
             else
                 return false;
         }
+        
+        // public bool GreaterThan(int a, int b) {
+        //     if (a > b)
+        //         return true;
+        //     else
+        //         return false;
+        // }
         
         // public bool GreaterThan(float a, float b) {
         //     if (a > b)
