+++ client-dotnet/test/EvoMaster.Instrumentation.Tests/Examples/Numbers/NumericOperationsTests.cs
@@ -14,7 +14,7 @@ namespace EvoMaster.Instrumentation.Tests.Examples.Numbers {
         [InlineData(8, 7, true)]
         [InlineData(4, 4400, false)]
         [InlineData(99, 99, false)]
-        public void TestGreaterThan(int a, int b, bool expectedResult) {
+        public void TestGreaterThan(double a, double b, bool expectedResult) {
             var actualResult = _numericOperations.GreaterThan(a, b);
 
             Assert.Equal(expectedResult, actualResult);
